model_evaluate_backend/requirements.txt
torch>=1.10.0
transformers>=4.20.0
tqdm>=4.64.0
numpy>=1.22.0
pyyaml>=6.0
openai>=1.0.0
requests>=2.27.0
pandas>=1.3.0 